app/requirements.txt
fastapi==0.115.12
pillow==11.1.0
protobuf==6.30.2
requests==2.32.3
transformers==4.50.3
google-cloud-vision==3.10.1
pydantic==2.11.1
pymongo==4.11.3
google-genai>=1.0.0
routes==2.5.1
uvicorn>=0.27.0  # Servidor ASGI             # Para validación de datos y modelos Pydantic
